+++ dbt/adapters/databricks/connections.py
@@ -35,6 +35,7 @@ from dbt.adapters.contracts.connection import DEFAULT_QUERY_COMMENT
 from dbt.adapters.contracts.connection import Identifier
 from dbt.adapters.contracts.connection import LazyHandle
 from dbt.adapters.databricks.__version__ import version as __version__
+from dbt.adapters.databricks.api_client import DatabricksApiClient
 from dbt.adapters.databricks.auth import BearerAuth
 from dbt.adapters.databricks.credentials import DatabricksCredentials
 from dbt.adapters.databricks.credentials import TCredentialProvider
@@ -477,16 +478,15 @@ class DatabricksConnectionManager(SparkConnectionManager):
     credentials_provider: Optional[TCredentialProvider] = None
     _user_agent = f"dbt-databricks/{__version__}"
 
+    def __init__(self, profile: AdapterRequiredConfig, mp_context: SpawnContext) -> None:
+        super().__init__(profile, mp_context)
+
     def cancel_open(self) -> List[str]:
         cancelled = super().cancel_open()
-        if self.credentials_provider:
-            logger.info("Cancelling open python jobs")
-            tracker = PythonRunTracker()
-            session = Session()
-            creds = self.credentials_provider(None)  # type: ignore
-            session.auth = BearerAuth(creds)
-            session.headers = {"User-Agent": self._user_agent}
-            tracker.cancel_runs(session)
+        creds = cast(DatabricksCredentials, self.profile.credentials)
+        api_client = DatabricksApiClient.create(creds, 15 * 60)
+        logger.info("Cancelling open python jobs")
+        PythonRunTracker.cancel_runs(api_client)
         return cancelled
 
     def compare_dbr_version(self, major: int, minor: int) -> int:
