+++ tests/unit/python/test_python_run_tracker.py
@@ -5,13 +5,17 @@ from mock import Mock
 class TestPythonRunTracker:
     def test_cancel_runs__from_separate_instance(self):
         tracker = PythonRunTracker()
-        tracker.set_host("host")
         tracker.insert_run_id("run_id")
-        other_tracker = PythonRunTracker()
-        mock_session = Mock()
+        mock_client = Mock()
 
-        other_tracker.cancel_runs(mock_session)
-        mock_session.post.assert_called_once_with(
-            "https://host/api/2.1/jobs/runs/cancel",
-            json={"run_id": "run_id"},
-        )
+        PythonRunTracker.cancel_runs(mock_client)
+        mock_client.job_runs.cancel.assert_called_once_with("run_id")
+
+    def test_cancel_runs__with_command_execution(self):
+        tracker = PythonRunTracker()
+        mock_command = Mock()
+        tracker.insert_command(mock_command)
+        mock_client = Mock()
+
+        PythonRunTracker.cancel_runs(mock_client)
+        mock_client.commands.cancel.assert_called_once_with(mock_command)
